pyproject.toml
[project]
name = "api-template"
version = "1.0.0"
requires-python = ">=3.9.6"

dependencies = [
  "flask>=3.1.2",
  "sqlalchemy>=2.0.45",
  "requests>=2.32.5",
  "python-dotenv>=1.2.1",
  "cx-Oracle>=8.2.0",
  "ruff>=0.14.7",
  "pre-commit>=4.3.0",
  "isort>=6.1.0",
  "pipreqs>=0.4.13",
  "uv>=0.9.18",
  "pymongo>=4.15.5",
  "sendgrid>=6.12.5",
  "openpyxl>=3.1.5",
]

# Exclude specific directories
exclude = [
    ".bzr",
    ".direnv", 
    ".eggs",
    ".git",
    ".venv",
    "__pycache__",
    "node_modules",
]

[tool.ruff.lint]
# Enable specific rule sets
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "I",      # isort
    "N",      # pep8-naming
    "D",      # pydocstyle
    "S",      # flake8-bandit
    "T20",    # flake8-print
    "PT",     # flake8-pytest-style
]

# Ignore specific Ruff rules
ignore = [
    "B008",  # Do not perform function calls in argument defaults
    "D100",  # Missing docstring in public module
    "D101",  # Missing docstring in public class
    "D102",  # Missing docstring in public method
    "D103",  # Missing docstring in public function
    "D104",  # Missing docstring in public package
    "D105",  # Missing docstring in magic method
    "D107",  # Missing docstring in __init__
    "D203",  # Conflicts with formatter (no blank line before class docstring)
    "D205",  # Blank line required between summary and description
    "D211",  # Conflicts with formatter (no blank line after class docstring)
    "D213",  # Conflicts with formatter (summary on second line)
    "D401",  # First line of docstring should be in imperative mood
    "D417",  # Missing argument descriptions in docstring
    "E402",  # Module name not at top of file
    "E501",  # Line too long
    "E722",  # Do not use bare except
    "I001",  # Import block is un-sorted or un-formatted
    "N999",  # Invalid module name (should be snake_case)
    "S104",  # Possible binding to all network interfaces (0.0.0.0)
    "S110",  # try-except-pass detected
    "T201",  # print statement found
]

# Allow fixing specific rules
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Indent with spaces
indent-style = "space"
# Use trailing commas
skip-magic-trailing-comma = false
# Auto-detect line ending
line-ending = "auto"
